session updated for apis

+++ app.py
@@ -178,11 +178,16 @@ def admin_logout():
 @app.route('/admin/users')
 def admin_users():
     user_session_id = request.cookies.get("user_session_id")
+    if user_session_id not in sessions:
+        return redirect("/admin")
     user = sessions[user_session_id]
     return render_template("admin/admin_dashboard.html",section="users",user=user)
 
 @app.route('/api/users', methods=["GET"])
 def user_list():
+    user_session_id = request.cookies.get("user_session_id")
+    if user_session_id not in sessions:
+        return redirect("/admin")
     db = get_database()
     users_collection = db.users
     users = list(users_collection.find({}))
@@ -192,6 +197,9 @@ def user_list():
 
 @app.route('/api/users/<user_id>', methods=["DELETE"])
 def api_delete_user(user_id):
+    user_session_id = request.cookies.get("user_session_id")
+    if user_session_id not in sessions:
+        return redirect("/admin")
     db = get_database()
     users_collection = db.users
 
@@ -207,11 +215,16 @@ def api_delete_user(user_id):
 @app.route('/admin/dishes')
 def admin_dishes():
     user_session_id = request.cookies.get("user_session_id")
+    if user_session_id not in sessions:
+        return redirect("/admin")
     user = sessions.get(user_session_id)
     return render_template("admin/admin_dashboard.html", section="dishes", user=user)
 
 @app.route('/api/dishes', methods=['POST'])
 def add_dish():
+    user_session_id = request.cookies.get("user_session_id")
+    if user_session_id not in sessions:
+        return redirect("/admin")
     db = get_database()
     dishes = db.dishes
 
@@ -264,6 +277,9 @@ def get_dishes():
 
 @app.route('/api/dishes/<id>', methods=['DELETE'])
 def delete_dish(id):
+    user_session_id = request.cookies.get("user_session_id")
+    if user_session_id not in sessions:
+        return redirect("/admin")
     db = get_database()
     dishes_collection = db.dishes
 
@@ -296,6 +312,9 @@ def get_single_dish(id):
 
 @app.route('/api/dishes/<id>', methods=['PUT'])
 def update_dish(id):
+    user_session_id = request.cookies.get("user_session_id")
+    if user_session_id not in sessions:
+        return redirect("/admin")
     db = get_database()
     dishes = db.dishes
 
@@ -336,11 +355,16 @@ def update_dish(id):
 @app.route('/admin/menu')
 def admin_menu():
     user_session_id = request.cookies.get("user_session_id")
+    if user_session_id not in sessions:
+        return redirect("/admin")
     user = sessions.get(user_session_id)
     return render_template("admin/admin_dashboard.html", section="menu", user=user)
 
 @app.route("/api/menu", methods=["POST"])
 def add_menu_dish():
+    user_session_id = request.cookies.get("user_session_id")
+    if user_session_id not in sessions:
+        return redirect("/admin")
     db = get_database()
     data = request.get_json()
     day = data["day"]
@@ -372,10 +396,18 @@ def get_menu_day(day):
 # Delete dish from day
 @app.route("/api/menu/<id>", methods=["DELETE"])
 def delete_menu_dish(id):
+    user_session_id = request.cookies.get("user_session_id")
+    if user_session_id not in sessions:
+        return redirect("/admin")
     db = get_database()
     db.menu.delete_one({"_id": ObjectId(id)})
     return jsonify({"success": True})
 
+
+# Route for customer dishes section
+@app.route("/dishes")
+def get_dish_page():
+    return render_template("dish.html")
 
 if __name__ == "__main__":
     app.run(debug=True)
